Fix delayed call of plot `_on_close()`
This event doesn't need connecting/disconnecting

As `Signal.update_plot()` has in-built checks, it is easier to simply
always have it connected.

Clean up plotting

Re-add direct call

Removed stray code line

Typos
