+++ hyperspy/drawing/image.py
@@ -276,12 +276,6 @@ class ImagePlot(BlittedFigure):
 
         self.connect()
 
-    def add_marker(self, marker):
-        marker.ax = self.ax
-        if marker.axes_manager is None:
-            marker.axes_manager = self.axes_manager
-        self.ax_markers.append(marker)
-
     def update(self, **kwargs):
         ims = self.ax.images
         # update extent:
@@ -436,14 +430,3 @@ class ImagePlot(BlittedFigure):
         while check_tolerance() and i <= step_prec_max:
             optimize_for_oom(step_oom - i)
             i += 1
-
-    def _on_close(self):
-        for marker in self.ax_markers:
-            marker.close()
-        self.events.closed.trigger(obj=self)
-        for f in self.events.closed.connected:
-            self.events.closed.disconnect(f)
-        self.figure = None
-
-    def close(self):
-        plt.close(self.figure)  # This will trigger self._on_close()
