+++ hyperspy/drawing/mpl_hse.py
@@ -118,18 +118,13 @@ class MPL_HyperSignal1D_Explorer(MPL_HyperExplorer):
         self.signal_plot = sf
         sf.plot()
         if self.navigator_plot is not None and sf.figure is not None:
-            utils.on_figure_window_close(self.navigator_plot.figure,
-                                         self._on_navigator_plot_closing)
-            utils.on_figure_window_close(sf.figure,
-                                         self.close_navigator_plot)
-            self._key_nav_cid = \
-                self.signal_plot.figure.canvas.mpl_connect(
-                    'key_press_event',
-                    self.axes_manager.key_navigator)
-            self._key_nav_cid = \
-                self.navigator_plot.figure.canvas.mpl_connect(
-                    'key_press_event',
-                    self.axes_manager.key_navigator)
+            self.navigator_plot.events.closed.connect(
+                self._on_navigator_plot_closing, [])
+            sf.events.closed.connect(self.close_navigator_plot, [])
+            self.signal_plot.figure.canvas.mpl_connect(
+                'key_press_event', self.axes_manager.key_navigator)
+            self.navigator_plot.figure.canvas.mpl_connect(
+                'key_press_event', self.axes_manager.key_navigator)
             self.signal_plot.figure.canvas.mpl_connect(
                 'key_press_event', self.key2switch_right_pointer)
             self.navigator_plot.figure.canvas.mpl_connect(
