+++ hyperspy/drawing/signal1d.py
@@ -100,12 +100,6 @@ class Signal1DFigure(BlittedFigure):
                 self._color_cycles[line.type].color_cycle.remove(
                     rgba_color)
 
-    def add_marker(self, marker):
-        marker.ax = self.ax
-        if marker.axes_manager is None:
-            marker.axes_manager = self.axes_manager
-        self.ax_markers.append(marker)
-
     def plot(self):
         self.ax.set_xlabel(self.xlabel)
         self.ax.set_ylabel(self.ylabel)
@@ -133,17 +127,11 @@ class Signal1DFigure(BlittedFigure):
                 pass
 
     def _on_close(self):
-        for marker in self.ax_markers:
-            marker.close()
+        if self.figure is None:
+            return  # Already closed
         for line in self.ax_lines + self.right_ax_lines:
             line.close()
-        self.events.closed.trigger(obj=self)
-        for f in self.events.closed.connected:
-            self.events.closed.disconnect(f)
-        self.figure = None
-
-    def close(self):
-        plt.close(self.figure)
+        super(Signal1DFigure, self)._on_close()
 
     def update(self):
         for marker in self.ax_markers:
